server3/requirements.txt
# Server 3 — Automation Layer Dependencies
# All pinned to major version. Install: pip install -r requirements.txt
# No CrewAI — Server 3 agents are standalone FastAPI endpoints.

fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
supabase>=2.4,<3.0
pydantic>=2.6,<3.0
httpx>=0.27,<1.0
python-multipart>=0.0.9,<1.0

# Job scraping
jobspy>=0.1,<1.0

# Browser automation
undetected-chromedriver>=3.5,<4.0
selenium>=4.18,<5.0
webdriver-manager>=4.0,<5.0

# Document generation
python-docx>=1.1,<2.0
PyPDF2>=3.0,<4.0
Pillow>=10.0,<12.0

# Custom scraping
beautifulsoup4>=4.12,<5.0

# Data processing
scikit-learn>=1.4,<2.0
numpy>=1.26,<3.0
python-dateutil>=2.9,<3.0
pytz>=2024.1

# Cryptography for AES session management
cryptography>=42.0,<44.0

# Compression (stdlib gzip used — no pip entry needed)
# All secrets via Doppler — no python-dotenv
